requirements.txt
beautifulsoup4
importlib-metadata
numpy
pem
packaging
pip>=21.3.1
pyOpenSSL>=21.0.0
pyyaml>=5.1
requests
statistics
stomp.py<=7.0.0
distro

opensearch-py
packaging
importlib_resources
# Not required to install the suite, included so it is added as a wheel for offline installation
git+https://gitlab.cern.ch/hep-benchmarks/hep-score.git@v1.5
